requirements.txt
feedparser==6.0.11
feedgen==0.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
